abstract nokogiri-related DS methods into a reusable mixin
